# docs/swagger.yaml
basePath: /
definitions:
  github_com_dgrco_quikslate_internal_domain.EmployeePosition:
    properties:
      position_id:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_dgrco_quikslate_internal_domain.Position:
    properties:
      business_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_dgrco_quikslate_internal_response.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_handler.MultipleEmployeePositionResponse:
    properties:
      positions:
        items:
          $ref: '#/definitions/github_com_dgrco_quikslate_internal_domain.EmployeePosition'
        type: array
    type: object
...
  internal_handler.MultiplePositionResponse:
    properties:
      positions:
        items:
          $ref: '#/definitions/github_com_dgrco_quikslate_internal_domain.Position'
        type: array
    type: object
...
  internal_handler.SimpleResponse:
    properties:
      message:
        type: string
    type: object
...
  internal_handler.SinglePositionResponse:
    properties:
      position:
        $ref: '#/definitions/github_com_dgrco_quikslate_internal_domain.Position'
    type: object
...
  internal_handler.addPositionRequest:
    properties:
      position_id:
        type: string
    type: object
...
  internal_handler.createPositionRequest:
    properties:
      name:
        type: string
    type: object
...
  internal_handler.renamePositionRequest:
    properties:
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: Scheduling API for businesses, locations, positions, shifts, and employees.
  title: QuikSlate API
  version: "1.0"
paths:
...
  /businesses/{businessId}/employees/{userId}/positions:
    get:
      description: List all positions assigned to a user within the business. Admin,
        Manager, or LocationLead (any location in the business).
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Target user ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MultipleEmployeePositionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: List employee positions
      tags:
      - employees
    post:
      consumes:
      - application/json
      description: Assign a position to a user. Admin, Manager, or LocationLead (any
        location in the business).
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Target user ID
        in: path
        name: userId
        required: true
        type: string
      - description: Position to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.addPositionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SimpleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Add employee position
      tags:
      - employees
  /businesses/{businessId}/employees/{userId}/positions/{positionId}:
    delete:
      description: Remove a position from a user. Admin only.
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Target user ID
        in: path
        name: userId
        required: true
        type: string
      - description: Position ID
        in: path
        name: positionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Remove employee position
      tags:
      - employees
...
  /businesses/{businessId}/positions:
    get:
      description: List all positions for a business. Business member only.
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MultiplePositionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: List positions
      tags:
      - positions
    post:
      consumes:
      - application/json
      description: Create a new position (job role, business-wide) under a business.
        Admin only.
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Position details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createPositionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SinglePositionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Create position
      tags:
      - positions
  /businesses/{businessId}/positions/{positionId}:
    delete:
      description: Delete a position. Admin only.
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Position ID
        in: path
        name: positionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete position
      tags:
      - positions
    get:
      description: Get a single position by ID. Business member only.
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Position ID
        in: path
        name: positionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SinglePositionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Get position
      tags:
      - positions
    patch:
      consumes:
      - application/json
      description: Rename a position. Admin only.
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Position ID
        in: path
        name: positionId
        required: true
        type: string
      - description: New name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.renamePositionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SimpleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Rename position
      tags:
      - positions
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"